refactor: combine overlay geom and retention into one
